coupon-system/api/src/test/java/com/rusty/api/service/ApplyServiceTest.java
package com.rusty.api.service;

import com.rusty.api.repository.CouponRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.concurrent.CountDownLatch;
import java.util.concurrent.ExecutorService;
import java.util.concurrent.Executors;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.api.Assertions.*;
@SpringBootTest
class ApplyServiceTest {

    @Autowired
    private ApplyService applyService;

    @Autowired
    private CouponRepository couponRepository;

    @Test
    void 한번만응모() {
//        applyService.apply(1L);
        applyService.apply(1L);

        long count = couponRepository.count();
        assertThat(count).isEqualTo(1);
    }

    @Test
    void 백번응모() throws InterruptedException {
        int threadCount = 10;
        ExecutorService executorService = Executors.newFixedThreadPool(32);
        CountDownLatch latch = new CountDownLatch(threadCount);

        for(int i = 0; i < threadCount; i++) {
            long userId = i;
            executorService.submit(() -> {
                try {
                    applyService.apply(userId);
                }finally {
                    latch.countDown();
                }
            });
        }

        latch.await();
        long count = couponRepository.count();
        assertThat(count).isEqualTo(100);

    }
}